+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Features
 
 - Add integration to collect Metrics, can be enabled by setting `options.enableMetrics = true` (#6956)
+- Add implementation for Metrics Protocol (#6960)
+
+### Fixes
+
+- Use `sendDefaultPii` and `span_id` for attributes in `BatcherScope` (#7055)
 
 ## 9.1.0